redesign CLI again
