+++ cmd/dr/config/create_dr.go
@@ -13,7 +13,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package dr
+package config
 
 import (
 	"fmt"
@@ -28,10 +28,10 @@ import (
 	ybmclient "github.com/yugabyte/yugabytedb-managed-go-client-internal"
 )
 
-var updateDrCmd = &cobra.Command{
-	Use:   "update",
-	Short: "Update DR for a cluster",
-	Long:  `Update DR for a cluster`,
+var createDrCmd = &cobra.Command{
+	Use:   "create",
+	Short: "Create DR for a cluster",
+	Long:  `Create DR for a cluster`,
 	Run: func(cmd *cobra.Command, args []string) {
 		authApi, err := ybmAuthClient.NewAuthApiClient()
 		if err != nil {
@@ -39,18 +39,19 @@ var updateDrCmd = &cobra.Command{
 		}
 		authApi.GetInfo("", "")
 
-		drName, _ := cmd.Flags().GetString("dr-name")
+		drName, _ := cmd.Flags().GetString("name")
+		sourceClusterName, _ := cmd.Flags().GetString("source-cluster")
+		targetClusterName, _ := cmd.Flags().GetString("target-cluster")
 		databases, _ := cmd.Flags().GetStringArray("databases")
+		sourceClusterId, err := authApi.GetClusterIdByName(sourceClusterName)
 		if err != nil {
 			logrus.Fatalf("Could not get cluster data: %s", ybmAuthClient.GetApiErrorDetails(err))
 		}
-		drInfo, err := authApi.GetDrDetailsByName(drName)
+		targetClusterId, err := authApi.GetClusterIdByName(targetClusterName)
 		if err != nil {
-			logrus.Fatal(err)
+			logrus.Fatalf("Could not get cluster data: %s", ybmAuthClient.GetApiErrorDetails(err))
 		}
-		drId := drInfo.GetId()
-		clusterId := drInfo.GetSourceClusterId()
-		namespacesResp, r, err := authApi.GetClusterNamespaces(clusterId).Execute()
+		namespacesResp, r, err := authApi.GetClusterNamespaces(sourceClusterId).Execute()
 		if err != nil {
 			logrus.Debugf("Full HTTP response: %v", r)
 			logrus.Fatalf(ybmAuthClient.GetApiErrorDetails(err))
@@ -69,27 +70,27 @@ var updateDrCmd = &cobra.Command{
 				}
 			}
 		}
-
-		updateDrRequest := ybmclient.NewEditXClusterDrRequest(*ybmclient.NewEditXClusterDrSpec(databaseIds))
-		drResp, response, err := authApi.EditXClusterDr(clusterId, drId).EditXClusterDrRequest(*updateDrRequest).Execute()
+		createDrRequest := ybmclient.NewCreateXClusterDrRequest(*ybmclient.NewXClusterDrSpec(drName, targetClusterId, databaseIds))
+		drResp, response, err := authApi.CreateXClusterDr(sourceClusterId).CreateXClusterDrRequest(*createDrRequest).Execute()
 		if err != nil {
 			logrus.Debugf("Full HTTP response: %v", response)
 			logrus.Fatalf(ybmAuthClient.GetApiErrorDetails(err))
 		}
+		drId := drResp.GetData().Info.Id
 
-		msg := fmt.Sprintf("DR config %s is being updated", formatter.Colorize(drName, formatter.GREEN_COLOR))
+		msg := fmt.Sprintf("The DR %s is being created", formatter.Colorize(drName, formatter.GREEN_COLOR))
 
 		if viper.GetBool("wait") {
-			returnStatus, err := authApi.WaitForTaskCompletion(clusterId, ybmclient.ENTITYTYPEENUM_CLUSTER, ybmclient.TASKTYPEENUM_EDIT_DR, []string{"FAILED", "SUCCEEDED"}, msg)
+			returnStatus, err := authApi.WaitForTaskCompletion(sourceClusterId, ybmclient.ENTITYTYPEENUM_CLUSTER, ybmclient.TASKTYPEENUM_CREATE_DR, []string{"FAILED", "SUCCEEDED"}, msg)
 			if err != nil {
 				logrus.Fatalf("error when getting task status: %s", err)
 			}
 			if returnStatus != "SUCCEEDED" {
 				logrus.Fatalf("Operation failed with error: %s", returnStatus)
 			}
-			fmt.Printf("DR config %s has been updated\n", formatter.Colorize(drName, formatter.GREEN_COLOR))
+			fmt.Printf("The DR %s has been created\n", formatter.Colorize(drName, formatter.GREEN_COLOR))
 
-			drGetResp, r, err := authApi.GetXClusterDr(clusterId, drId).Execute()
+			drGetResp, r, err := authApi.GetXClusterDr(sourceClusterId, drId).Execute()
 			if err != nil {
 				logrus.Debugf("Full HTTP response: %v", r)
 				logrus.Fatalf(ybmAuthClient.GetApiErrorDetails(err))
@@ -110,9 +111,13 @@ var updateDrCmd = &cobra.Command{
 }
 
 func init() {
-	DrCmd.AddCommand(updateDrCmd)
-	updateDrCmd.Flags().String("dr-name", "", "[REQUIRED] Name of the DR configuration.")
-	updateDrCmd.MarkFlagRequired("dr-name")
-	updateDrCmd.Flags().StringArray("databases", []string{}, "[REQUIRED] Databases to be replicated. Please provide a comma separated list of database names <db-name-1>,<db-name-2>.")
-	updateDrCmd.MarkFlagRequired("databases")
+	ConfigCmd.AddCommand(createDrCmd)
+	createDrCmd.Flags().String("name", "", "[REQUIRED] Name of the DR configuration.")
+	createDrCmd.MarkFlagRequired("name")
+	createDrCmd.Flags().String("source-cluster", "", "[REQUIRED] Source cluster in the DR configuration.")
+	createDrCmd.MarkFlagRequired("source-cluster")
+	createDrCmd.Flags().String("target-cluster", "", "[REQUIRED] Target cluster in the DR configuration.")
+	createDrCmd.MarkFlagRequired("target-cluster")
+	createDrCmd.Flags().StringArray("databases", []string{}, "[REQUIRED] Databases to be replicated. Please provide a comma separated list of database names <db-name-1>,<db-name-2>.")
+	createDrCmd.MarkFlagRequired("databases")
 }
