+++ cmd/dr/config/delete_dr.go
@@ -13,23 +13,33 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package dr
+package config
 
 import (
-	"os"
+	"fmt"
 
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"github.com/yugabyte/ybm-cli/cmd/util"
 	ybmAuthClient "github.com/yugabyte/ybm-cli/internal/client"
 	"github.com/yugabyte/ybm-cli/internal/formatter"
 	ybmclient "github.com/yugabyte/yugabytedb-managed-go-client-internal"
 )
 
-var describeDrCmd = &cobra.Command{
-	Use:   "describe",
-	Short: "Describe DR",
-	Long:  `Describe DR`,
+var deleteDrCmd = &cobra.Command{
+	Use:   "delete",
+	Short: "Delete DR",
+	Long:  `Delete DR`,
+	PreRun: func(cmd *cobra.Command, args []string) {
+		viper.BindPFlag("force", cmd.Flags().Lookup("force"))
+		drName, _ := cmd.Flags().GetString("config")
+		msg := fmt.Sprintf("Are you sure you want to delete dr: %s", drName)
+		err := util.ConfirmCommand(msg, viper.GetBool("force"))
+		if err != nil {
+			logrus.Fatal(err)
+		}
+	},
 	Run: func(cmd *cobra.Command, args []string) {
 		authApi, err := ybmAuthClient.NewAuthApiClient()
 		if err != nil {
@@ -37,9 +47,9 @@ var describeDrCmd = &cobra.Command{
 		}
 		authApi.GetInfo("", "")
 
-		drName, _ := cmd.Flags().GetString("dr-name")
+		drName, _ := cmd.Flags().GetString("config")
 		if err != nil {
-			logrus.Fatalf("Could not get cluster data: %s", ybmAuthClient.GetApiErrorDetails(err))
+			logrus.Fatal(err)
 		}
 		drInfo, err := authApi.GetDrDetailsByName(drName)
 		if err != nil {
@@ -47,23 +57,33 @@ var describeDrCmd = &cobra.Command{
 		}
 		drId := drInfo.GetId()
 		clusterId := drInfo.GetSourceClusterId()
-		drResp, r, err := authApi.GetXClusterDr(clusterId, drId).Execute()
+
+		r, err := authApi.DeleteXClusterDr(clusterId, drId).Execute()
 		if err != nil {
 			logrus.Debugf("Full HTTP response: %v", r)
 			logrus.Fatalf(ybmAuthClient.GetApiErrorDetails(err))
 		}
 
-		drCtx := formatter.Context{
-			Output: os.Stdout,
-			Format: formatter.NewDrFormat(viper.GetString("output")),
-		}
+		msg := fmt.Sprintf("The DR %s is being deleted", formatter.Colorize(drName, formatter.GREEN_COLOR))
 
-		formatter.DrWrite(drCtx, []ybmclient.XClusterDrData{drResp.GetData()}, *authApi)
+		if viper.GetBool("wait") {
+			returnStatus, err := authApi.WaitForTaskCompletion(clusterId, ybmclient.ENTITYTYPEENUM_CLUSTER, ybmclient.TASKTYPEENUM_DELETE_DR, []string{"FAILED", "SUCCEEDED"}, msg)
+			if err != nil {
+				logrus.Fatalf("error when getting task status: %s", err)
+			}
+			if returnStatus != "SUCCEEDED" {
+				logrus.Fatalf("Operation failed with error: %s", returnStatus)
+			}
+			fmt.Printf("The DR %s has been deleted\n", formatter.Colorize(drName, formatter.GREEN_COLOR))
+			return
+		}
+		fmt.Println(msg)
 	},
 }
 
 func init() {
-	DrCmd.AddCommand(describeDrCmd)
-	describeDrCmd.Flags().String("dr-name", "", "[REQUIRED] Name of the DR configuration.")
-	describeDrCmd.MarkFlagRequired("dr-name")
+	ConfigCmd.AddCommand(deleteDrCmd)
+	deleteDrCmd.Flags().String("config", "", "[REQUIRED] Name of the DR configuration")
+	deleteDrCmd.MarkFlagRequired("config")
+
 }
