+++ cmd/dr/config/describe_dr.go
@@ -13,33 +13,23 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package dr
+package config
 
 import (
-	"fmt"
+	"os"
 
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"github.com/yugabyte/ybm-cli/cmd/util"
 	ybmAuthClient "github.com/yugabyte/ybm-cli/internal/client"
 	"github.com/yugabyte/ybm-cli/internal/formatter"
 	ybmclient "github.com/yugabyte/yugabytedb-managed-go-client-internal"
 )
 
-var deleteDrCmd = &cobra.Command{
-	Use:   "delete",
-	Short: "Delete DR",
-	Long:  `Delete DR`,
-	PreRun: func(cmd *cobra.Command, args []string) {
-		viper.BindPFlag("force", cmd.Flags().Lookup("force"))
-		drName, _ := cmd.Flags().GetString("dr-name")
-		msg := fmt.Sprintf("Are you sure you want to delete dr: %s", drName)
-		err := util.ConfirmCommand(msg, viper.GetBool("force"))
-		if err != nil {
-			logrus.Fatal(err)
-		}
-	},
+var describeDrCmd = &cobra.Command{
+	Use:   "describe",
+	Short: "Describe DR",
+	Long:  `Describe DR`,
 	Run: func(cmd *cobra.Command, args []string) {
 		authApi, err := ybmAuthClient.NewAuthApiClient()
 		if err != nil {
@@ -47,9 +37,9 @@ var deleteDrCmd = &cobra.Command{
 		}
 		authApi.GetInfo("", "")
 
-		drName, _ := cmd.Flags().GetString("dr-name")
+		drName, _ := cmd.Flags().GetString("config")
 		if err != nil {
-			logrus.Fatal(err)
+			logrus.Fatalf("Could not get cluster data: %s", ybmAuthClient.GetApiErrorDetails(err))
 		}
 		drInfo, err := authApi.GetDrDetailsByName(drName)
 		if err != nil {
@@ -57,33 +47,23 @@ var deleteDrCmd = &cobra.Command{
 		}
 		drId := drInfo.GetId()
 		clusterId := drInfo.GetSourceClusterId()
-
-		r, err := authApi.DeleteXClusterDr(clusterId, drId).Execute()
+		drResp, r, err := authApi.GetXClusterDr(clusterId, drId).Execute()
 		if err != nil {
 			logrus.Debugf("Full HTTP response: %v", r)
 			logrus.Fatalf(ybmAuthClient.GetApiErrorDetails(err))
 		}
 
-		msg := fmt.Sprintf("The DR %s is being deleted", formatter.Colorize(drName, formatter.GREEN_COLOR))
-
-		if viper.GetBool("wait") {
-			returnStatus, err := authApi.WaitForTaskCompletion(clusterId, ybmclient.ENTITYTYPEENUM_CLUSTER, ybmclient.TASKTYPEENUM_DELETE_DR, []string{"FAILED", "SUCCEEDED"}, msg)
-			if err != nil {
-				logrus.Fatalf("error when getting task status: %s", err)
-			}
-			if returnStatus != "SUCCEEDED" {
-				logrus.Fatalf("Operation failed with error: %s", returnStatus)
-			}
-			fmt.Printf("The DR %s has been deleted\n", formatter.Colorize(drName, formatter.GREEN_COLOR))
-			return
+		drCtx := formatter.Context{
+			Output: os.Stdout,
+			Format: formatter.NewDrFormat(viper.GetString("output")),
 		}
-		fmt.Println(msg)
+
+		formatter.DrWrite(drCtx, []ybmclient.XClusterDrData{drResp.GetData()}, *authApi)
 	},
 }
 
 func init() {
-	DrCmd.AddCommand(deleteDrCmd)
-	deleteDrCmd.Flags().String("dr-name", "", "[REQUIRED] Name of the DR configuration")
-	deleteDrCmd.MarkFlagRequired("dr-name")
-
+	ConfigCmd.AddCommand(describeDrCmd)
+	describeDrCmd.Flags().String("config", "", "[REQUIRED] Name of the DR configuration.")
+	describeDrCmd.MarkFlagRequired("config")
 }
