+++ cmd/root.go
@@ -43,7 +43,6 @@ import (
 	"github.com/yugabyte/ybm-cli/cmd/vpc"
 
 	"github.com/yugabyte/ybm-cli/internal/log"
-	"github.com/yugabyte/ybm-cli/internal/releases"
 )
 
 var (
@@ -66,7 +65,7 @@ var rootCmd = &cobra.Command{
 		if strings.HasPrefix(cmd.CommandPath(), "ybm completion") {
 			return
 		}
-		releases.PrintUpgradeMessageIfNeeded()
+		//releases.PrintUpgradeMessageIfNeeded()
 
 	},
 }
